+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+0.4.2b
+---
+* Improved cache and get/load data system
+
 0.4.1b
 ---
 
 * Fixed bug that shows unfollowers IDs in stdout
 * Fixed bug that made crash the software at the first run
 * Fixed issue: saving after showing unfollowers names, now before
-* Fixed issue: Unfollowers removed like non-followers
+* Fixed issue: Unfollowers removed like non-followers (0.4.1b2)
 
 0.4b
 ---